requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.1.0
pydantic==2.11.4
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
langchain==0.3.25
langchain-community==0.3.24
langchain-groq==0.3.2
faiss-cpu==1.10.0
sentence-transformers==4.1.0
PyPDF2==3.0.1
python-docx==1.1.2
httpx==0.28.1
langsmith==0.3.42
